+++ epyk/core/html/HtmlInput.py
@@ -1250,7 +1250,7 @@ class Search(Html.Html):
                      'vertical-align': 'top'})
     else:
       self.input.css({"text-align": 'left', 'padding-left': "2px", 'padding-right': '%spx' % Defaults.LINE_HEIGHT})
-      self.icon.css({"margin": '2px 5px 10px 0px', 'cursor': 'pointer', "right": 0,
+      self.icon.css({"margin": '5px 5px 5px 0px', 'cursor': 'pointer', "right": 0,
                      'position': 'absolute', 'vertical-align': 'top'})
     self.tooltip(tooltip)